Returning to my mathematical roots.
I'm going to be doing some ML work now. I'm tired of getting rejected
for frontend jobs because of my weak portfolio when I don't think that
I would like the jobs in the first place. So I'm going to be doing a lot
of math, and a lot of ml fun.

Here is the corrected probability sampler. It can create uniform, single
and multivariable normal, categorical, and mixture model samples from
distributions with a specified mean and (co)variance. the plotsample.py is
a script that I wrote with plotting functions so that I could test my models
by eye.

I use the Box Muller transform and the Cholesky decomposition for normal samples.