Scrapping precomputation
A friend and colleague in the UC Berkeley Symphony is a ML/NLP
grad student at Cal. I asked him a little bit, and he said that
I just needed to split my data on relevant attributes to narrow
the search space. I did that and am now running my cross validation
function to try and find the right k for k-nearest neighbors.