Working on speeding up cross validation
At the expense of a small possibility of reducing true randomness
I'm implementing a partial lookup for validation, that means that
given a partition there will be memoization potential. Since my partition
function can be made to return the same partitions over and over
for each value of k in the k-nearest-neighbors evaluation. Right now I'm
debugging a couple of type errors. I need to just write some tests and
I should be good though.