Decision Trees
So I wanted to implement an algorithm to construct decision trees.
It eats an importance function for splitting variables during construction,
and is built to handle both continuous and discrete attributes.

Now I want to implement a random forest algorithm so that I can
investigate variable correleation for the purpose of feature extraction.